agentic-system/requirements.txt
# Core dependencies
crewai>=0.102.0
crewai-tools>=0.36.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Web search and API tools
tavily-python>=0.5.1
firecrawl>=1.12.0
urlextract>=1.9.0

# Google API integration
google-auth>=2.38.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.161.0

# Graph RAG dependencies
langchain>=0.3.19
langchain-openai>=0.2.14
langchain-neo4j>=0.2.0
langchain-core>=0.3.40
langchain-community>=0.3.18
langchain-experimental>=0.3.4
langchain-text-splitters>=0.3.6
neo4j>=5.27.0
openai>=1.65.2

# Scheduling and logging
schedule>=1.2.0
logging>=0.5.1

# Utility packages
datetime
pathlib
pickle
re
